requirements.txt
tensorflow-cpu==2.13.0
numpy==1.26.4
Pillow==10.3.0
